+++ src/NodaTime/Fields/DateTimeFieldBase.cs
@@ -14,6 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 #endregion
+
 using System;
 
 namespace NodaTime.Fields
@@ -24,7 +25,7 @@ namespace NodaTime.Fields
     public abstract class DateTimeFieldBase : IDateTimeField
     {
         private readonly DateTimeFieldType fieldType;
-        
+
         protected DateTimeFieldBase(DateTimeFieldType fieldType)
         {
             if (fieldType == null)
@@ -76,7 +77,6 @@ namespace NodaTime.Fields
         public abstract bool IsLenient { get; }
 
         #region Values
-
         /// <summary>
         /// Get the value of this field from the local instant.
         /// </summary>
@@ -169,7 +169,7 @@ namespace NodaTime.Fields
         public virtual LocalInstant AddWrapField(LocalInstant localInstant, int value)
         {
             int current = GetValue(localInstant);
-            int wrapped = FieldUtils.GetWrappedValue(current, value, (int)GetMinimumValue(localInstant), (int)GetMaximumValue(localInstant));
+            int wrapped = FieldUtils.GetWrappedValue(current, value, (int) GetMinimumValue(localInstant), (int) GetMaximumValue(localInstant));
             return SetValue(localInstant, wrapped);
         }
 
@@ -220,7 +220,7 @@ namespace NodaTime.Fields
                 long proposed = values[fieldIndex] + valueToAdd;
                 if (proposed <= max)
                 {
-                    values[fieldIndex] = (int)proposed;
+                    values[fieldIndex] = (int) proposed;
                     break;
                 }
                 if (nextField == null)
@@ -236,9 +236,9 @@ namespace NodaTime.Fields
                         throw new ArgumentException("Fields invalid for add");
                     }
                 }
-                valueToAdd -= ((int)max + 1) - values[fieldIndex];  // reduce the amount to add
-                values = nextField.Add(instant, fieldIndex - 1, values, 1);  // add 1 to next bigger field
-                values[fieldIndex] = (int)GetMinimumValue(instant, values);  // reset this field to zero
+                valueToAdd -= ((int) max + 1) - values[fieldIndex]; // reduce the amount to add
+                values = nextField.Add(instant, fieldIndex - 1, values, 1); // add 1 to next bigger field
+                values[fieldIndex] = (int) GetMinimumValue(instant, values); // reset this field to zero
             }
             while (valueToAdd < 0)
             {
@@ -246,7 +246,7 @@ namespace NodaTime.Fields
                 long proposed = values[fieldIndex] + valueToAdd;
                 if (proposed >= min)
                 {
-                    values[fieldIndex] = (int)proposed;
+                    values[fieldIndex] = (int) proposed;
                     break;
                 }
                 if (nextField == null)
@@ -261,12 +261,12 @@ namespace NodaTime.Fields
                         throw new ArgumentException("Fields invalid for add");
                     }
                 }
-                valueToAdd -= ((int)min - 1) - values[fieldIndex];  // reduce the amount to add
-                values = nextField.Add(instant, fieldIndex - 1, values, -1);  // subtract 1 from next bigger field
-                values[fieldIndex] = (int)GetMaximumValue(instant, values);  // reset this field to max value
+                valueToAdd -= ((int) min - 1) - values[fieldIndex]; // reduce the amount to add
+                values = nextField.Add(instant, fieldIndex - 1, values, -1); // subtract 1 from next bigger field
+                values[fieldIndex] = (int) GetMaximumValue(instant, values); // reset this field to max value
             }
 
-            return SetValue(instant, fieldIndex, values, values[fieldIndex]);  // adjusts smaller fields
+            return SetValue(instant, fieldIndex, values, values[fieldIndex]); // adjusts smaller fields
         }
 
         /// <summary>
@@ -312,11 +312,11 @@ namespace NodaTime.Fields
             IDateTimeField nextField = null;
             while (valueToAdd > 0)
             {
-                int max = (int)GetMaximumValue(instant, values);
+                int max = (int) GetMaximumValue(instant, values);
                 long proposed = values[fieldIndex] + valueToAdd;
                 if (proposed <= max)
                 {
-                    values[fieldIndex] = (int)proposed;
+                    values[fieldIndex] = (int) proposed;
                     break;
                 }
                 if (nextField == null)
@@ -324,7 +324,7 @@ namespace NodaTime.Fields
                     if (fieldIndex == 0)
                     {
                         valueToAdd -= (max + 1) - values[fieldIndex];
-                        values[fieldIndex] = (int)GetMinimumValue(instant, values);
+                        values[fieldIndex] = (int) GetMinimumValue(instant, values);
                         continue;
                     }
                     nextField = instant.GetField(fieldIndex - 1);
@@ -334,17 +334,17 @@ namespace NodaTime.Fields
                         throw new ArgumentException("Fields invalid for add");
                     }
                 }
-                valueToAdd -= (max + 1) - values[fieldIndex];  // reduce the amount to add
-                values = nextField.AddWrapPartial(instant, fieldIndex - 1, values, 1);  // add 1 to next bigger field
-                values[fieldIndex] = (int)GetMinimumValue(instant, values);  // reset this field to zero
+                valueToAdd -= (max + 1) - values[fieldIndex]; // reduce the amount to add
+                values = nextField.AddWrapPartial(instant, fieldIndex - 1, values, 1); // add 1 to next bigger field
+                values[fieldIndex] = (int) GetMinimumValue(instant, values); // reset this field to zero
             }
             while (valueToAdd < 0)
             {
-                int min = (int)GetMinimumValue(instant, values);
+                int min = (int) GetMinimumValue(instant, values);
                 long proposed = values[fieldIndex] + valueToAdd;
                 if (proposed >= min)
                 {
-                    values[fieldIndex] = (int)proposed;
+                    values[fieldIndex] = (int) proposed;
                     break;
                 }
                 if (nextField == null)
@@ -352,7 +352,7 @@ namespace NodaTime.Fields
                     if (fieldIndex == 0)
                     {
                         valueToAdd -= (min - 1) - values[fieldIndex];
-                        values[fieldIndex] = (int)GetMaximumValue(instant, values);
+                        values[fieldIndex] = (int) GetMaximumValue(instant, values);
                         continue;
                     }
                     nextField = instant.GetField(fieldIndex - 1);
@@ -361,12 +361,12 @@ namespace NodaTime.Fields
                         throw new ArgumentException("Fields invalid for add");
                     }
                 }
-                valueToAdd -= ((int)min - 1) - values[fieldIndex];  // reduce the amount to add
-                values = nextField.AddWrapPartial(instant, fieldIndex - 1, values, -1);  // subtract 1 from next bigger field
-                values[fieldIndex] = (int)GetMaximumValue(instant, values);  // reset this field to max value
+                valueToAdd -= (min - 1) - values[fieldIndex]; // reduce the amount to add
+                values = nextField.AddWrapPartial(instant, fieldIndex - 1, values, -1); // subtract 1 from next bigger field
+                values[fieldIndex] = (int) GetMaximumValue(instant, values); // reset this field to max value
             }
 
-            return SetValue(instant, fieldIndex, values, values[fieldIndex]);  // adjusts smaller fields
+            return SetValue(instant, fieldIndex, values, values[fieldIndex]); // adjusts smaller fields
         }
 
         /// <summary>
@@ -492,11 +492,11 @@ namespace NodaTime.Fields
                 IDateTimeField field = instant.GetField(i);
                 if (values[i] > field.GetMaximumValue(instant, values))
                 {
-                    values[i] = (int)field.GetMaximumValue(instant, values);
+                    values[i] = (int) field.GetMaximumValue(instant, values);
                 }
-                if (values[i] < (int)field.GetMinimumValue(instant, values))
+                if (values[i] < (int) field.GetMinimumValue(instant, values))
                 {
-                    values[i] = (int)field.GetMinimumValue(instant, values);
+                    values[i] = (int) field.GetMinimumValue(instant, values);
                 }
             }
             return values;
@@ -516,11 +516,9 @@ namespace NodaTime.Fields
             int value = ConvertText(text, provider);
             return SetValue(instant, fieldIndex, values, value);
         }
-
         #endregion
 
         #region Leap
-
         /// <summary>
         /// Defaults to non-leap.
         /// </summary>
@@ -541,11 +539,9 @@ namespace NodaTime.Fields
         /// Defaults to null, i.e. no leap duration field.
         /// </summary>
         public virtual IDurationField LeapDurationField { get { return null; } }
-
         #endregion
 
         #region Ranges
-
         /// <summary>
         /// Defaults to the absolute maximum for the field.
         /// </summary>
@@ -615,11 +611,9 @@ namespace NodaTime.Fields
         /// </summary>
         /// <returns>The minimum valid value for this field, in the units of the field</returns>
         public abstract long GetMinimumValue();
-
         #endregion
 
         #region Rounding
-
         /// <summary>
         /// Round to the lowest whole unit of this field. After rounding, the value
         /// of this field and all fields of a higher magnitude are retained. The
@@ -672,7 +666,7 @@ namespace NodaTime.Fields
             Duration diffFromFloor = localInstant - floor;
             Duration diffToCeiling = ceiling - localInstant;
 
-             // Closer to the floor, or halfway - round floor
+            // Closer to the floor, or halfway - round floor
             return diffFromFloor <= diffToCeiling ? floor : ceiling;
         }
 
@@ -692,7 +686,7 @@ namespace NodaTime.Fields
             long diffFromFloor = localInstant.Ticks - floor.Ticks;
             long diffToCeiling = ceiling.Ticks - localInstant.Ticks;
 
-             // Closer to the ceiling, or halfway - round ceiling
+            // Closer to the ceiling, or halfway - round ceiling
             return diffToCeiling <= diffFromFloor ? ceiling : floor;
         }
 
@@ -717,7 +711,7 @@ namespace NodaTime.Fields
             {
                 return floor;
             }
-            // Closer to the ceiling - round ceiling
+                // Closer to the ceiling - round ceiling
             else if (diffToCeiling < diffFromFloor)
             {
                 return ceiling;
@@ -744,11 +738,9 @@ namespace NodaTime.Fields
         {
             return localInstant - RoundFloor(localInstant);
         }
-
         #endregion
 
         #region Text
-
         /// <summary>
         /// Get the maximum text value for this field.
         /// </summary>
@@ -760,7 +752,7 @@ namespace NodaTime.Fields
         /// </remarks>
         public int GetMaximumTextLength(IFormatProvider provider)
         {
-            int max = (int)GetMaximumValue();
+            int max = (int) GetMaximumValue();
             if (max >= 0)
             {
                 if (max < 10)
@@ -941,7 +933,6 @@ namespace NodaTime.Fields
         {
             return GetAsText(fieldValue, provider);
         }
-
         #endregion
 
         protected int ConvertText(String text, IFormatProvider provider)
